feat: finish implementing first hacky version of additional users backend

+++ app/Http/Controllers/AdditionalUserController.php
@@ -6,6 +6,7 @@ use App\Models\AdditionalUser;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Session;
 
 class AdditionalUserController extends Controller
 {
@@ -64,18 +65,60 @@ class AdditionalUserController extends Controller
         $loggedInUserId = Auth::id();
 
         $validatedRequest = $request->validate([
-            'additional_user_id' => ['required', 'exists:users,id']
+            'additional_user_email' => ['required', 'exists:users,email']
         ]);
 
+        $additionalUserData = User::where('email', $validatedRequest['additional_user_email'])->get()->first();
+
         $additionalUser = AdditionalUser::where('user_id', $loggedInUserId)
-            ->where('additional_user_id', $validatedRequest['additional_user_id'])
+            ->where('additional_user_id', $additionalUserData->id)
             ->get()
             ->first();
+
+        if (!$additionalUser) {
+            return response([
+                'errors' => ['Email address is not an additional user of your account.']
+            ], 400);
+        }
 
         $additionalUser->delete();
 
         return [
             'message' => 'Successfully removed additional user.'
+        ];
+    }
+
+    public function loginAsAnotherUser(Request $request)
+    {
+        $loggedInUserId = Auth::id();
+
+        $validatedRequest = $request->validate([
+            'user_email_to_login_as' => ['required', 'exists:users,email']
+        ]);
+
+        $userToLoginAs = User::where('email', $validatedRequest['user_email_to_login_as'])->get()->first();
+
+        $additionalUserEntry = AdditionalUser::where('user_id', $userToLoginAs->id)
+            ->where('additional_user_id', $loggedInUserId)
+            ->get()
+            ->first();
+
+        if (!$additionalUserEntry) {
+            return response([
+                'errors' => ['You do not have access to this account.']
+            ], 400);
+        }
+
+        Auth::logout();
+        $request->session()->invalidate();
+        $request->session()->regenerateToken();
+
+        Auth::loginUsingId($userToLoginAs->id);
+
+        Session::put('additional_user_id', $loggedInUserId);
+
+        return [
+            'message' => 'Successfully logged into account as additional user.'
         ];
     }
 }
